+++ internal/app/unit/converter_test.go
@@ -8,9 +8,9 @@ import (
 )
 
 func TestGigaBytesToBytes(t *testing.T) {
-	assert.Equal(t, int64(1073741824), converter.GigaBytesToBytes(int32(1)), "1 GB conversion is not correct")
+	assert.Equal(t, int32(1073741824), converter.GigaBytesToBytes(int32(1)), "1 GB conversion is not correct")
 	assert.Equal(t, int64(1073741824), converter.GigaBytesToBytes(int64(1)), "1 GB conversion is not correct")
-	assert.Equal(t, int64(1073741824), converter.GigaBytesToBytes(float64(1)), "1 GB conversion is not correct")
+	assert.Equal(t, float64(1073741824), converter.GigaBytesToBytes(float64(1)), "1 GB conversion is not correct")
 }
 
 func TestMegaBytesToBytes(t *testing.T) {
